test(tui): align sidebar rail tests with the package layout
